+++ delimited/__init__.py
@@ -1,9 +1,14 @@
+"""
+delimited
+~~~~~~~~~
+"""
+
 from delimited.type import NestedType
 from delimited.path import TuplePath
 from delimited.path import DelimitedStrPath
 
 
-__version__ = "0.0.7"
+__version__ = "0.0.8"
 __title__ = "delimited"
 __summary__ = "Defines types that allow for accessing and modifying nested data"
 __url__ = "https://github.com/chrisantonellis/delimited"